fix: pasting error in table when using custom sanitizer (fix #980)
