+++ apps/editor/src/js/htmlSanitizer.js
@@ -61,7 +61,7 @@ function htmlSanitizer(html, needHtmlText) {
   leaveOnlyWhitelistAttribute(root);
   removeInvalidAttributeValues(root);
 
-  return finalizeHtml(root, needHtmlText);
+  return domUtils.finalizeHtml(root, needHtmlText);
 }
 
 /**
@@ -124,32 +124,6 @@ function removeInvalidAttributeValues(html) {
       });
     }
   }
-}
-
-/**
- * Finalize html result
- * @private
- * @param {HTMLElement} html root element
- * @param {boolean} needHtmlText pass true if need html text
- * @returns {string|DocumentFragment} result
- */
-function finalizeHtml(html, needHtmlText) {
-  let returnValue;
-
-  if (needHtmlText) {
-    returnValue = html.innerHTML;
-  } else {
-    const frag = document.createDocumentFragment();
-    const childNodes = toArray(html.childNodes);
-    const { length } = childNodes;
-
-    for (let i = 0; i < length; i += 1) {
-      frag.appendChild(childNodes[i]);
-    }
-    returnValue = frag;
-  }
-
-  return returnValue;
 }
 
 export default htmlSanitizer;
